Fix - [Character] SetITTMovementMode<T> 인수 Cast 누락 수정 #01d8aa2

+++ Source/ITT/Component/Character/Movement/ITTCharacterMovementComponent.h
@@ -44,7 +44,7 @@ public:
 	template<typename T>
 	void SetITTMovementMode(EMovementMode MainMode,  T SubMode, uint8 AdditiveMode1 = 0, uint8 AdditiveMode2 = 0, bool bChangeImmediately = true)
 	{
-		SetITTMovementMode(static_cast<uint8>(MainMode), SubMode, AdditiveMode1, AdditiveMode2, bChangeImmediately);
+		SetITTMovementMode(static_cast<uint8>(MainMode), static_cast<uint8>(SubMode), AdditiveMode1, AdditiveMode2, bChangeImmediately);
 	}
